+++ go/api/base_client.go
@@ -6754,8 +6754,8 @@ func (client *baseClient) GeoDistWithUnit(
 //		searchBy - The query's shape options:
 //		 - `BYRADIUS` to search inside circular area according to given radius.
 //		 - `BYBOX` to search inside an axis-aligned rectangle, determined by height and width.
-//		geoSearchResultOptions - Optional inputs for sorting/limiting the results.
-//		geoSearchOptions - The optional inputs to request additional information.
+//		resultOptions - Optional inputs for sorting/limiting the results.
+//		infoOptions - The optional inputs to request additional information.
 //
 // Return value:
 //
@@ -6771,8 +6771,8 @@ func (client *baseClient) GeoSearchWithFullOptions(
 	key string,
 	searchFrom options.GeoSearchOrigin,
 	searchByShape options.GeoSearchShape,
-	geoSearchResultOptions options.GeoSearchResultOptions,
-	geoSearchOptions options.GeoSearchOptions,
+	resultOptions options.GeoSearchResultOptions,
+	infoOptions options.GeoSearchInfoOptions,
 ) ([]any, error) {
 	args := []string{key}
 	searchFromArgs, err := searchFrom.ToArgs()
@@ -6785,16 +6785,16 @@ func (client *baseClient) GeoSearchWithFullOptions(
 		return nil, err
 	}
 	args = append(args, searchByShapeArgs...)
-	geoSearchOptionsArgs, err := geoSearchOptions.ToArgs()
+	infoOptionsArgs, err := infoOptions.ToArgs()
 	if err != nil {
 		return nil, err
 	}
-	args = append(args, geoSearchOptionsArgs...)
-	geoSearchResultOptionsArgs, err := geoSearchResultOptions.ToArgs()
+	args = append(args, infoOptionsArgs...)
+	resultOptionsArgs, err := resultOptions.ToArgs()
 	if err != nil {
 		return nil, err
 	}
-	args = append(args, geoSearchResultOptionsArgs...)
+	args = append(args, resultOptionsArgs...)
 	result, err := client.executeCommand(C.GeoSearch, args)
 	if err != nil {
 		return nil, err
@@ -6832,7 +6832,7 @@ func (client *baseClient) GeoSearchWithResultOptions(
 	key string,
 	searchFrom options.GeoSearchOrigin,
 	searchByShape options.GeoSearchShape,
-	geoSearchResultOptions options.GeoSearchResultOptions,
+	resultOptions options.GeoSearchResultOptions,
 ) ([]string, error) {
 	args := []string{key}
 	searchFromArgs, err := searchFrom.ToArgs()
@@ -6845,14 +6845,12 @@ func (client *baseClient) GeoSearchWithResultOptions(
 		return nil, err
 	}
 	args = append(args, searchByShapeArgs...)
-	geoSearchResultOptionsArgs, err := geoSearchResultOptions.ToArgs()
+	resultOptionsArgs, err := resultOptions.ToArgs()
 	if err != nil {
 		return nil, err
 	}
-	args = append(args, geoSearchResultOptionsArgs...)
-	// fmt.Println("================================================")
-	// fmt.Println("args:", args)
-	// fmt.Println("================================================")
+	args = append(args, resultOptionsArgs...)
+
 	result, err := client.executeCommand(C.GeoSearch, args)
 	if err != nil {
 		return nil, err
@@ -6891,18 +6889,18 @@ func (client *baseClient) GeoSearchWithResultOptions(
 //	 - The coordinates as a two item `array` of `float64`.
 //
 // [valkey.io]: https://valkey.io/commands/geosearch/
-func (client *baseClient) GeoSearchWithOptions(
+func (client *baseClient) GeoSearchWithInfoOptions(
 	key string,
 	searchFrom options.GeoSearchOrigin,
 	searchByShape options.GeoSearchShape,
-	geoSearchOptions options.GeoSearchOptions,
+	infoOptions options.GeoSearchInfoOptions,
 ) ([]any, error) {
 	return client.GeoSearchWithFullOptions(
 		key,
 		searchFrom,
 		searchByShape,
 		*options.NewGeoSearchResultOptions(),
-		geoSearchOptions,
+		infoOptions,
 	)
 }
 
@@ -6941,5 +6939,216 @@ func (client *baseClient) GeoSearch(
 		searchFrom,
 		searchByShape,
 		*options.NewGeoSearchResultOptions(),
+	)
+}
+
+// Searches for members in a sorted set stored at `sourceKey` representing geospatial data
+// within a circular or rectangular area and stores the result in `destinationKey`. If
+// `destinationKey` already exists, it is overwritten. Otherwise, a new sorted set will be
+// created. To get the result directly, see [GeoSearchWithFullOptions].
+//
+// Since:
+//
+//	Valkey 6.2.0 and above.
+//
+// Note: When in cluster mode, `destination` and all `keys` must map to the same hash slot.
+//
+// See [valkey.io] for more details.
+//
+// Parameters:
+//
+//	destinationKey - The key of the sorted set to store the result.
+//	sourceKey - The key of the sorted set to search.
+//	searchFrom - The query's center point options, could be one of:
+//		 - `MemberOrigin` to use the position of the given existing member in the sorted
+//	          set.
+//		 - `CoordOrigin` to use the given longitude and latitude coordinates.
+//		searchBy - The query's shape options:
+//		 - `BYRADIUS` to search inside circular area according to given radius.
+//		 - `BYBOX` to search inside an axis-aligned rectangle, determined by height and width.
+//		resultOptions - Optional inputs for sorting/limiting the results.
+//		infoOptions - The optional inputs to request additional information.
+//
+// Return value:
+//
+//	The number of elements in the resulting set.
+//
+// [valkey.io]: https://valkey.io/commands/geosearchstore/
+func (client *baseClient) GeoSearchStoreWithFullOptions(
+	destinationKey string,
+	sourceKey string,
+	searchFrom options.GeoSearchOrigin,
+	searchByShape options.GeoSearchShape,
+	resultOptions options.GeoSearchResultOptions,
+	storeInfoOptions options.GeoSearchStoreInfoOptions,
+) (int64, error) {
+	args := []string{destinationKey, sourceKey}
+	searchFromArgs, err := searchFrom.ToArgs()
+	if err != nil {
+		return defaultIntResponse, err
+	}
+	args = append(args, searchFromArgs...)
+	searchByShapeArgs, err := searchByShape.ToArgs()
+	if err != nil {
+		return defaultIntResponse, err
+	}
+	args = append(args, searchByShapeArgs...)
+	resultOptionsArgs, err := resultOptions.ToArgs()
+	if err != nil {
+		return defaultIntResponse, err
+	}
+	args = append(args, resultOptionsArgs...)
+	storeInfoOptionsArgs, err := storeInfoOptions.ToArgs()
+	if err != nil {
+		return defaultIntResponse, err
+	}
+	args = append(args, storeInfoOptionsArgs...)
+
+	result, err := client.executeCommand(C.GeoSearchStore, args)
+	if err != nil {
+		return defaultIntResponse, err
+	}
+	return handleIntResponse(result)
+}
+
+// Searches for members in a sorted set stored at `sourceKey` representing geospatial data
+// within a circular or rectangular area and stores the result in `destinationKey`. If
+// `destinationKey` already exists, it is overwritten. Otherwise, a new sorted set will be
+// created. To get the result directly, see [GeoSearchWithFullOptions].
+//
+// Since:
+//
+//	Valkey 6.2.0 and above.
+//
+// Note: When in cluster mode, `destination` and all `keys` must map to the same hash slot.
+//
+// See [valkey.io] for more details.
+//
+// Parameters:
+//
+//	destinationKey - The key of the sorted set to store the result.
+//	sourceKey - The key of the sorted set to search.
+//	searchFrom - The query's center point options, could be one of:
+//		 - `MemberOrigin` to use the position of the given existing member in the sorted
+//	          set.
+//		 - `CoordOrigin` to use the given longitude and latitude coordinates.
+//		searchBy - The query's shape options:
+//		 - `BYRADIUS` to search inside circular area according to given radius.
+//		 - `BYBOX` to search inside an axis-aligned rectangle, determined by height and width.
+//
+// Return value:
+//
+//	The number of elements in the resulting set.
+//
+// [valkey.io]: https://valkey.io/commands/geosearchstore/
+func (client *baseClient) GeoSearchStore(
+	destinationKey string,
+	sourceKey string,
+	searchFrom options.GeoSearchOrigin,
+	searchByShape options.GeoSearchShape,
+) (int64, error) {
+	return client.GeoSearchStoreWithFullOptions(
+		destinationKey,
+		sourceKey,
+		searchFrom,
+		searchByShape,
+		*options.NewGeoSearchResultOptions(),
+		*options.NewGeoSearchStoreInfoOptions(),
+	)
+}
+
+// Searches for members in a sorted set stored at `sourceKey` representing geospatial data
+// within a circular or rectangular area and stores the result in `destinationKey`. If
+// `destinationKey` already exists, it is overwritten. Otherwise, a new sorted set will be
+// created. To get the result directly, see [GeoSearchWithFullOptions].
+//
+// Since:
+//
+//	Valkey 6.2.0 and above.
+//
+// Note: When in cluster mode, `destination` and all `keys` must map to the same hash slot.
+//
+// See [valkey.io] for more details.
+//
+// Parameters:
+//
+//	destinationKey - The key of the sorted set to store the result.
+//	sourceKey - The key of the sorted set to search.
+//	searchFrom - The query's center point options, could be one of:
+//		 - `MemberOrigin` to use the position of the given existing member in the sorted
+//	          set.
+//		 - `CoordOrigin` to use the given longitude and latitude coordinates.
+//		searchBy - The query's shape options:
+//		 - `BYRADIUS` to search inside circular area according to given radius.
+//		 - `BYBOX` to search inside an axis-aligned rectangle, determined by height and width.
+//		resultOptions - Optional inputs for sorting/limiting the results.
+//
+// Return value:
+//
+//	The number of elements in the resulting set.
+//
+// [valkey.io]: https://valkey.io/commands/geosearchstore/
+func (client *baseClient) GeoSearchStoreWithResultOptions(
+	destinationKey string,
+	sourceKey string,
+	searchFrom options.GeoSearchOrigin,
+	searchByShape options.GeoSearchShape,
+	resultOptions options.GeoSearchResultOptions,
+) (int64, error) {
+	return client.GeoSearchStoreWithFullOptions(
+		destinationKey,
+		sourceKey,
+		searchFrom,
+		searchByShape,
+		resultOptions,
+		*options.NewGeoSearchStoreInfoOptions(),
+	)
+}
+
+// Searches for members in a sorted set stored at `sourceKey` representing geospatial data
+// within a circular or rectangular area and stores the result in `destinationKey`. If
+// `destinationKey` already exists, it is overwritten. Otherwise, a new sorted set will be
+// created. To get the result directly, see [GeoSearchWithFullOptions].
+//
+// Since:
+//
+//	Valkey 6.2.0 and above.
+//
+// Note: When in cluster mode, `destination` and all `keys` must map to the same hash slot.
+//
+// See [valkey.io] for more details.
+//
+// Parameters:
+//
+//	destinationKey - The key of the sorted set to store the result.
+//	sourceKey - The key of the sorted set to search.
+//	searchFrom - The query's center point options, could be one of:
+//		 - `MemberOrigin` to use the position of the given existing member in the sorted
+//	          set.
+//		 - `CoordOrigin` to use the given longitude and latitude coordinates.
+//		searchBy - The query's shape options:
+//		 - `BYRADIUS` to search inside circular area according to given radius.
+//		 - `BYBOX` to search inside an axis-aligned rectangle, determined by height and width.
+//		infoOptions - The optional inputs to request additional information.
+//
+// Return value:
+//
+//	The number of elements in the resulting set.
+//
+// [valkey.io]: https://valkey.io/commands/geosearchstore/
+func (client *baseClient) GeoSearchStoreWithInfoOptions(
+	destinationKey string,
+	sourceKey string,
+	searchFrom options.GeoSearchOrigin,
+	searchByShape options.GeoSearchShape,
+	infoOptions options.GeoSearchStoreInfoOptions,
+) (int64, error) {
+	return client.GeoSearchStoreWithFullOptions(
+		destinationKey,
+		sourceKey,
+		searchFrom,
+		searchByShape,
+		*options.NewGeoSearchResultOptions(),
+		infoOptions,
 	)
 }