+++ go/api/geospacial_commands.go
@@ -35,6 +35,13 @@ type GeoSpatialCommands interface {
 		searchByShape options.GeoSearchShape,
 	) ([]string, error)
 
+	GeoSearchWithInfoOptions(
+		key string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		infoOptions options.GeoSearchInfoOptions,
+	) ([]any, error)
+
 	GeoSearchWithResultOptions(
 		key string,
 		searchFrom options.GeoSearchOrigin,
@@ -42,18 +49,43 @@ type GeoSpatialCommands interface {
 		resultOptions options.GeoSearchResultOptions,
 	) ([]string, error)
 
-	GeoSearchWithOptions(
-		key string,
-		searchFrom options.GeoSearchOrigin,
-		searchByShape options.GeoSearchShape,
-		options options.GeoSearchOptions,
-	) ([]any, error)
-
 	GeoSearchWithFullOptions(
 		key string,
 		searchFrom options.GeoSearchOrigin,
 		searchByShape options.GeoSearchShape,
 		resultOptions options.GeoSearchResultOptions,
-		options options.GeoSearchOptions,
+		infoOptions options.GeoSearchInfoOptions,
 	) ([]any, error)
+
+	GeoSearchStore(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+	) (int64, error)
+
+	GeoSearchStoreWithInfoOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		storeInfoOptions options.GeoSearchStoreInfoOptions,
+	) (int64, error)
+
+	GeoSearchStoreWithResultOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+	) (int64, error)
+
+	GeoSearchStoreWithFullOptions(
+		destinationKey string,
+		sourceKey string,
+		searchFrom options.GeoSearchOrigin,
+		searchByShape options.GeoSearchShape,
+		resultOptions options.GeoSearchResultOptions,
+		storeInfoOptions options.GeoSearchStoreInfoOptions,
+	) (int64, error)
 }