+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Go/Core: Move FFI to a dedicated folder for reusability ([#3372](https://github.com/valkey-io/valkey-glide/pull/3372))
 * Go: Add `GeoPos` ([#3409](https://github.com/valkey-io/valkey-glide/pull/3409))
 * Go: Add `GeoDist` ([#3446](https://github.com/valkey-io/valkey-glide/pull/3446))
+* Go: Add `GeoSearch` and `GeoSearchStore` ([#3385](https://github.com/valkey-io/valkey-glide/pull/3385))
 
 #### Breaking Changes